==> Scripts/08b_caseType.R <==
library(tidyverse)

require(zoo)
require(data.table)

##############  How to plot factors in a specified order in ggplot

############################  https://rstudio-pubs-static.s3.amazonaws.com/7433_4537ea5073dc4162950abb715f513469.html



read.aantal.landelijk.path <- paste("C:\\Rdir\\data\\",Sys.Date()-1,"\\", Sys.Date()-1, "_COVID-19_casus_landelijk.csv",sep="")
cases_per_yesterday <- read.csv(read.aantal.landelijk.path,sep=";")

read.aantal.landelijk.path <- paste("C:\\Rdir\\data\\",Sys.Date(),"\\", Sys.Date(), "_COVID-19_casus_landelijk.csv",sep="")
cases_per_day <- read.csv(read.aantal.landelijk.path,sep=";")

cases_per_yesterday$Date_statistics_type[cases_per_yesterday$Date_statistics_type == "DOO"] <- "DOO_old"
cases_per_yesterday$Date_statistics_type[cases_per_yesterday$Date_statistics_type == "DON"] <- "DON_old"
cases_per_yesterday$Date_statistics_type[cases_per_yesterday$Date_statistics_type == "DPL"] <- "DPL_old"


cases_per_yesterday$value <- 1
cases_per_yesterday_long <- aggregate(cases_per_yesterday$value, by = list(Type_Datum = cases_per_yesterday$Date_statistics_type, Datum = cases_per_yesterday$date, Dag = cases_per_yesterday$Date_file), FUN = sum)
colnames(cases_per_yesterday_long)[4] <- "value"
cases_per_yesterday_long <- cases_per_yesterday_long[,c("Datum", "Type_Datum", "value" )]


cases_per_day$value <- 1
cases_per_day_long <- aggregate(cases_per_day$value, by = list(Type_Datum = cases_per_day$Date_statistics_type, Datum = cases_per_day$date, Dag = cases_per_day$Date_file), FUN = sum)
colnames(cases_per_day_long)[4] <- "value"
cases_per_day_long <- cases_per_day_long[,c("Datum", "Type_Datum", "value" )]


cases_per_day_long_bind <- rbind(cases_per_day_long, cases_per_yesterday_long)   # Adding Rows (adding other df, to the bottom)
cases_per_day_long_spread_diff <- spread(cases_per_day_long_bind, Type_Datum, value)
cases_per_day_long_spread_diff[is.na(cases_per_day_long_spread_diff)] <- 0


#cases_per_day_long_spread_diff$DON_diff <- cases_per_day_long_spread_diff$DON - cases_per_day_long_spread_diff$DON_old
cases_per_day_long_spread_diff$DOO_diff <- cases_per_day_long_spread_diff$DOO - cases_per_day_long_spread_diff$DOO_old
cases_per_day_long_spread_diff$DPL_diff <- cases_per_day_long_spread_diff$DPL - cases_per_day_long_spread_diff$DPL_old

#cases_per_day_long_spread_diff$DON_diff_neg <- cases_per_day_long_spread_diff$DON_diff
#cases_per_day_long_spread_diff$DON_diff_neg[cases_per_day_long_spread_diff$DON_diff_neg > 0] <- 0
#cases_per_day_long_spread_diff$DON_old = cases_per_day_long_spread_diff$DON_old + cases_per_day_long_spread_diff$DON_diff_neg

cases_per_day_long_spread_diff$DOO_diff_neg <- cases_per_day_long_spread_diff$DOO_diff
cases_per_day_long_spread_diff$DOO_diff_neg[cases_per_day_long_spread_diff$DOO_diff_neg > 0] <- 0
cases_per_day_long_spread_diff$DOO_old = cases_per_day_long_spread_diff$DOO_old + cases_per_day_long_spread_diff$DOO_diff_neg

cases_per_day_long_spread_diff$DPL_diff_neg <- cases_per_day_long_spread_diff$DPL_diff
cases_per_day_long_spread_diff$DPL_diff_neg[cases_per_day_long_spread_diff$DPL_diff_neg > 0] <- 0
cases_per_day_long_spread_diff$DPL_old = cases_per_day_long_spread_diff$DPL_old + cases_per_day_long_spread_diff$DPL_diff_neg


cases_per_day_long_spread_diff <- cases_per_day_long_spread_diff[,c("Datum", "DOO_old","DOO_diff", "DPL_old", "DPL_diff" )]


#### Gather ####

keycol <- "Datum"
valuecol <- "type"
gathercols <- c("DOO_old","DOO_diff", "DPL_old", "DPL_diff")

cases_per_day_diff_gather<- gather(cases_per_day_long_spread_diff, keycol, valuecol, gathercols)
cases_per_day_diff_gather$Datum <- as.Date(cases_per_day_diff_gather$Datum)



### press events                         

dates_vline <- data.frame(date=as.Date(c("2020-09-18", "2020-09-28", "2020-10-13", "2020-11-05", 
                                         "2020-11-18", "2020-12-15", "2021-01-23", "2021-02-08",  "2021-03-31",
                                         "2021-06-26", "2021-07-09", "2021-08-30","2021-09-25",
                                         "2021-11-02", "2021-11-12", "2021-11-28", "2021-12-19",
                                         "2022-01-15")),
                          event=c("kroeg uurtje eerder dicht (regio)","R van 0,9 (landelijke maatregelen)","semi-lockdown",
                                  "verzwaring", "einde verzwaring", "lockdown", "                    avondklok",
                                 "basisscholen open", "avondklok 22:00uur",
                                 "Alles open!", "aanscherpingen", "geen 1,5meter in HO", "geen 1,5 meter meer",
                                 "Halve maatregelen", "Harde klap", "Avondclockdown", "lockdown",
                                 "Einde lockdown"))


df4 <- cases_per_day_diff_gather[cases_per_day_diff_gather$Datum>"2020-07-01",]





ggplot(df4, aes(x=Datum, y=valuecol, fill = factor(keycol, levels=c("DOO_diff","DOO_old","DPL_diff","DPL_old")), width=.7)) +
 geom_col()+

theme_classic()+
    xlab("")+ 
    ylab("")+

  
  scale_x_date(date_breaks = "month", 
               date_labels= format("%b"),
               limits = as.Date(c("2020-07-01", Sys.Date() ))
               )+

    scale_fill_manual(values=c(
                               "#548235", 
                               "#C5E0B4",  
                               "#203864",   #blauw donker
                               "#B4C7E7"),  #blauw
                      labels=c(   
                                  "Eerste ziektedag (nieuw/correctie)",
                                  "Eerste ziektedag",
                                  "Positieve labuitslag (nieuw/correctie)",
                                  "Positieve labuitslag"))+
  
  
  geom_vline(data=dates_vline,  mapping=aes(xintercept=date),  col = "darkgray", lwd = 0.5, linetype= "dashed")+
  geom_text(data=dates_vline  , mapping=aes(x=date, y=14000, label=event), size=5, angle=90, vjust=-0.4, hjust="right", alpha = 0.8)+
  
  scale_y_continuous(labels = label_comma(big.mark = ".", decimal.mark = ","))+
  
labs(title = "Besmette personen: verschil met gisteren",
     caption = paste("Bron: RIVM | Plot: @YorickB ",Sys.Date()))+
  
   theme(legend.position = c(0.07, 0.4),
        legend.background = element_rect(fill="#F5F5F5",size=0.8,linetype="solid",colour ="black"),
        legend.title = element_blank(),
        legend.text = element_text(colour="black", size=8, face="bold"))+
  
  theme(
    plot.background = element_rect(fill = "#F5F5F5"), #background color/size (border color and size)
    panel.background = element_rect(fill = "#F5F5F5", colour = "#F5F5F5"),
    plot.title = element_text(hjust = 0.5,size = 30,face = "bold"),
    plot.subtitle =  element_text(hjust=0.5 ,size = 15,color = "black", face = "italic"),
    axis.text.x = element_text(face="bold", color="black", size=12), #, angle=45),
    axis.text.y = element_text(face="bold", color="black", size=14),
    axis.ticks = element_line(colour = "#F5F5F5", size = 1, linetype = "solid"),
    axis.ticks.length = unit(0.5, "cm"),
    axis.line = element_line(colour = "#F5F5F5"),
    

    panel.grid.major.y = element_line(colour= "lightgray", linetype = "dashed"))+



ggsave("data/07_cases_type1.png",width=16, height = 9)  







df4_short  <- cases_per_day_diff_gather[cases_per_day_diff_gather$Datum>"2021-06-01",]

ggplot(df4_short, aes(x=Datum, y=valuecol, fill = factor(keycol, levels=c("DOO_diff","DOO_old","DPL_diff","DPL_old")), width=.7)) +
  geom_col()+
  
  theme_classic()+
  xlab("")+ 
  ylab("")+
  
  geom_vline(data=dates_vline,  mapping=aes(xintercept=date),  col = "darkgray", lwd = 0.5, linetype= "dashed")+
  geom_text(data=dates_vline  , mapping=aes(x=date, y=4000, label=event), size=8, angle=90, vjust=-0.4, hjust=0)+
  
  scale_x_date(date_breaks = "14 day", 
               date_labels= format("%d-%b"),
               limits = as.Date(c("2021-06-01", Sys.Date()+5)))+

  
  scale_fill_manual(values=c(
                             "#548235", 
                             "#C5E0B4",  
                             "#203864",   #blauw donker
                             "#B4C7E7"),  #blauw
                    labels=c(   
                                "Eerste ziektedag (nieuw/correctie)",
                                "Eerste ziektedag",
                                "Positieve labuitslag (nieuw/correctie)",
                                "Positieve labuitslag"))+
  
  
  scale_y_continuous( labels = label_comma(big.mark = ".", decimal.mark = ","))+
 
  labs(title = "Besmette personen: verschil met gisteren",
      # subtitle = "P",
       caption = paste("Source: RIVM | Plot: @YorickB | ",Sys.Date()))+
  
  theme(legend.position = c(0.4, 0.8),
        legend.background = element_rect(fill="#F5F5F5",size=0.8,linetype="solid",colour ="black"),
        legend.title = element_blank(),
        legend.text = element_text(colour="black", size=10, face="bold"))+
  
  theme(
    plot.background = element_rect(fill = "#F5F5F5"),
    panel.background = element_rect(fill = "#F5F5F5", colour = "#F5F5F5"),
    plot.title = element_text(hjust = 0.5,size = 30,face = "bold"),
    plot.subtitle =  element_text(hjust=0.5,size = 18,color = "black", face = "italic"),
    axis.text.x = element_text(face="bold", color="black", size=12),
    axis.text.y = element_text(face="bold", color="black", size=14),
    axis.ticks = element_line(colour = "#F5F5F5", size = 1, linetype = "solid"),
    axis.ticks.length = unit(0.5, "cm"),
    axis.line = element_line(colour = "#F5F5F5"),
    panel.grid.major.y = element_line(colour= "lightgray", linetype = "dashed"))+ 

  ggsave("data/07_cases_type1_summer_2021.png",width=16, height = 9)  






##mondays
dates_vline_mondays <- as.Date(c("2020-08-10","2020-08-17","2020-08-24",
                  "2020-08-31",
                  "2020-09-07",
                   "2020-09-14",
                  "2020-09-21",
                  "2020-09-28",
                   "2020-10-05",
                   "2020-10-12",
                  "2020-10-19",
                  "2020-10-26",
                  "2020-11-02",
                  "2020-11-09",
                  "2020-11-16",
                  "2020-11-23",
                  "2020-11-30",
                  "2020-12-07",
                  "2020-12-14",
                  "2020-12-21",
                  "2020-12-28",
                  "2021-01-04",
                  "2021-01-11",
                  "2021-01-18",
                  "2021-01-25",
                  "2021-02-01",
                  "2021-02-08",
                  "2021-02-15",
                  "2021-02-22",
                  "2021-03-01",
                  "2021-03-08",
                  "2021-03-15",
                  "2021-03-22",
                  "2021-03-29",
                  "2021-04-05",
                  "2021-04-12",
                  "2021-04-19",
                  "2021-04-26",
                  "2021-05-03",
                  "2021-05-10",
                  "2021-05-17",
                  "2021-05-24",
                  "2021-05-31",
                  "2021-11-01",
                  "2021-11-08",
                  "2021-11-15",
                  "2021-11-22"
                  
                  
                 ))   

dates_vline_mondays <- which((cases_per_day_diff_gather$Datum %in% dates_vline_mondays))




ggplot(cases_per_day_diff_gather, aes(x=Datum, y=valuecol, fill = factor(keycol, levels=c("DOO_diff","DOO_old","DPL_diff","DPL_old")), width=.7)) +
  geom_col()+

  theme_classic()+
  xlab("")+ 
  ylab("")+
    scale_x_date(date_breaks = "month", 
               date_labels= format("%b"),
               limits = as.Date(c("2021-10-01", Sys.Date() ))
  )+
  
  scale_fill_manual(values=c("#548235", "#C5E0B4", "#203864", "#B4C7E7"), 
                    labels=c(    "Eerste ziektedag (nieuw/correctie)",
                                 "Eerste ziektedag",
                                 "Positieve labuitslag (nieuw/correctie)",
                                "Positieve labuitslag"
                                ))+
  
  # geom_vline(xintercept = as.numeric(df4$Datum[dates_vline_mondays]),
  #           col = "darkgray", lwd = 0.5, linetype= "dashed")+
    # geom_text(mapping=aes(x=date, y=0, label=event), size=4, angle=90, vjust=-0.4, hjust=0) +
    labs(title = "Besmette personen: verschil met gisteren",
    #   subtitle = # "Maandagen", #  OMT: 'Een lagere R is beter'",
       caption = paste("Bron: RIVM | Plot: @YorickB ",Sys.Date()))+
  
  theme(legend.position = c(0.2, 0.8),
        legend.background = element_rect(fill="#F5F5F5",size=0.8,linetype="solid",colour ="black"),
        legend.title = element_blank(),
        legend.text = element_text(colour="black", size=9, face="bold"))+
  theme(
    plot.background = element_rect(fill = "#F5F5F5"), #background color/size (border color and size)
    panel.background = element_rect(fill = "#F5F5F5", colour = "#F5F5F5"),
    plot.title = element_text(hjust = 0.5,size = 30,face = "bold"),
    plot.subtitle =  element_text(hjust=0.5 ,size = 15,color = "black", face = "italic"),
    axis.text.x = element_text(face="bold", color="black", size=12), #, angle=45),
    axis.text.y = element_text(face="bold", color="black", size=14),
    axis.ticks = element_line(colour = "#F5F5F5", size = 1, linetype = "solid"),
    axis.ticks.length = unit(0.5, "cm"),
    axis.line = element_line(colour = "#F5F5F5"),
        #axis.labels.x=date_format("%d-%b"),
        panel.grid.major.y = element_line(colour= "lightgray", linetype = "dashed"))+ #,
  #panel.grid.major.x = element_line(colour= "darkgray", linetype = "solid"))
    ggsave("data/plots/07_cases_type1-monday.png",width=16, height = 9)






#### Gather ####



key <- "Datum"
value <- "type"
gathercols <- c("DPL_diff","DOO_diff")

dat_wide <- gather(cases_per_day_long_spread_diff, key, value, all_of(gathercols))

dat_wide$Datum <- as.Date(dat_wide$Datum)

two.weeks.ago <-as.Date(Sys.Date()-15)
dat_wide <- dat_wide[dat_wide$Datum>two.weeks.ago,]



ggplot(dat_wide, aes(x=Datum, y=value, fill = factor(key, levels=c("DOO_diff","DPL_diff")), width=.7)) +
  geom_col(position = position_dodge(width = 0.65))+
  
  theme_classic()+
  
  xlab("")+ 
  ylab("")+
  
  scale_x_date(date_breaks = "1 day", 
               date_labels= format("%d-%b"),
               limits = as.Date(c(Sys.Date()-15, Sys.Date()+1)))+
  
  scale_fill_manual(values=c("#548235", "#2f5597"), labels=c( "Eerste ziektedag (nieuw/correctie)",
                                                                         "Positieve labuitslag (nieuw/correctie)"
  ))+
  
  labs(title = "Besmette personen: toegevoegd / gecorrigeerd",
       caption = paste("Bron: RIVM | Plot: @YorickB | ",Sys.Date()))+
  
  theme(legend.position = c(0.2, 0.9),
        legend.background = element_rect(fill="#F5F5F5",size=0.8,linetype="solid",colour ="black"),
        legend.title = element_blank(),
        legend.text = element_text(colour="black", size=10, face="bold"))+
  
  theme(
    plot.background = element_rect(fill = "#F5F5F5"),
    panel.background = element_rect(fill = "#F5F5F5", colour = "#F5F5F5"),
    plot.title = element_text(hjust = 0.5,size = 30,face = "bold"),
    axis.text.x = element_text(face="bold", color="black", size=12),
    axis.text.y = element_text(face="bold", color="black", size=14),
    axis.ticks = element_line(colour = "#F5F5F5", size = 1, linetype = "solid"),
    axis.ticks.length = unit(0.5, "cm"),
    axis.line = element_line(colour = "#F5F5F5"),
    panel.grid.major.y = element_line(colour= "lightgray", linetype = "dashed"))


ggsave("data/07_cases_diff.png",width=16, height = 9)
